Fix MDX review findings: i64 validation, uniform repeat rule
The independent review approved the reader with four LOW findings,
all addressed here. Size and count validation arithmetic now runs in
i64 because the previous int products could wrap on a 32-bit target
and let a malformed count reach an allocation; wasm is a standing
portability constraint, so this is fixed now rather than deferred
to the port. Duplicate top-level chunk detection previously keyed on
the first chunk's parsed data being non-nil, so an empty first chunk
let a repeat through; a seen-tag set now enforces the rule uniformly
for all typed tags (unknown tags stay repeatable since skipping has
no side effects), which also made the repeated-MODL cleanup dance
unreachable and it is removed. Six new tests pin the previously
uncovered paths: the empty-then-repeat loophole, unknown-tag repeat
tolerance, interpolation type above Bezier, unknown layer sub-tag,
duplicated in-node track (keeps last, leak-free), and a node size
below the minimum.

Sweep unchanged after the fixes: 3362/3362 models parse.

+++ src/formats/mdx/mdx_tracks.odin
@@ -24,7 +24,7 @@ parse_track_set :: proc(reader: ^Reader, track_set: ^Track_Set($Value), allocato
 	}
 	has_tangents := interpolation_raw >= u32(Interpolation_Type.Hermite)
 	bytes_per_key := 4 + size_of(Value) * (3 if has_tangents else 1)
-	if int(key_count) * bytes_per_key > reader_remaining(reader) {
+	if i64(key_count) * i64(bytes_per_key) > i64(reader_remaining(reader)) {
 		return .Corrupt_Chunk
 	}
 	// A duplicated track tag within one object does not occur in the
